Add otel tracing for some critical paths
Minor rant, but why is OTEL so complicated for even the most simple of
use-cases? It has no reason to be this complicated. Every turn is a
deprecation or a surprise. There's SO much room for a package that
simplifies the damn thing or a competitor that actually makes tracing
easy and accessible to folks who don't have significant experience with
tracing. Raise the floor, not the ceiling.